chore: move header to its components
